Add tests for CSV loading
