--- src/test/java/mx/uam/archinaut/ApplicationTests.java
@@ -1,13 +1,0 @@
-package mx.uam.archinaut;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
