+++ src/test/java/mx/uam/archinaut/services/MetricsServiceTest.java
@@ -1,0 +1,43 @@
+package mx.uam.archinaut.services;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+import java.io.IOException;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import com.opencsv.exceptions.CsvValidationException;
+
+import mx.uam.archinaut.model.DesignStructureMatrix;
+import mx.uam.archinaut.model.MatrixElement;
+
+@SpringBootTest
+public class MetricsServiceTest extends AbstractServiceTest  {
+	
+	@Autowired
+	private DesignStructureMatrixService designStructureMatrixService;
+	
+	@Autowired
+	private MetricsService metricsService;
+	
+	@Test
+	void testLoadMetrics() throws CsvValidationException, IOException {
+		
+		DesignStructureMatrix matrix = designStructureMatrixService.loadMatrixFromJSON(dependsConfigurationEntry);
+		assertNotNull(matrix);
+		
+		matrix = metricsService.loadMetrics(nonDependsConfigurationEntries, matrix);
+		assertNotNull(matrix);
+		
+		for (MatrixElement m : matrix.getElements()) {
+						
+			System.out.println(m.getFullName());
+			System.out.println(m.getMetrics().size());
+			
+		}
+		
+	}
+
+}
